requirements.txt
paramiko=2.11.0
PyYAML=6.0